chore: clean up repos
