+++ server/src/services/map.service.spec.ts
@@ -1,5 +1,4 @@
 import { IAlbumRepository } from 'src/interfaces/album.interface';
-import { IAssetRepository } from 'src/interfaces/asset.interface';
 import { ILoggerRepository } from 'src/interfaces/logger.interface';
 import { IMapRepository } from 'src/interfaces/map.interface';
 import { IPartnerRepository } from 'src/interfaces/partner.interface';
@@ -8,7 +7,6 @@ import { MapService } from 'src/services/map.service';
 import { assetStub } from 'test/fixtures/asset.stub';
 import { authStub } from 'test/fixtures/auth.stub';
 import { newAlbumRepositoryMock } from 'test/repositories/album.repository.mock';
-import { newAssetRepositoryMock } from 'test/repositories/asset.repository.mock';
 import { newLoggerRepositoryMock } from 'test/repositories/logger.repository.mock';
 import { newMapRepositoryMock } from 'test/repositories/map.repository.mock';
 import { newPartnerRepositoryMock } from 'test/repositories/partner.repository.mock';
@@ -18,21 +16,19 @@ import { Mocked } from 'vitest';
 describe(MapService.name, () => {
   let sut: MapService;
   let albumMock: Mocked<IAlbumRepository>;
-  let assetMock: Mocked<IAssetRepository>;
   let loggerMock: Mocked<ILoggerRepository>;
   let partnerMock: Mocked<IPartnerRepository>;
   let mapMock: Mocked<IMapRepository>;
   let systemMetadataMock: Mocked<ISystemMetadataRepository>;
 
   beforeEach(() => {
-    assetMock = newAssetRepositoryMock();
     albumMock = newAlbumRepositoryMock();
     loggerMock = newLoggerRepositoryMock();
     partnerMock = newPartnerRepositoryMock();
     mapMock = newMapRepositoryMock();
     systemMetadataMock = newSystemMetadataRepositoryMock();
 
-    sut = new MapService(albumMock, assetMock, loggerMock, partnerMock, mapMock, systemMetadataMock);
+    sut = new MapService(albumMock, loggerMock, partnerMock, mapMock, systemMetadataMock);
   });
 
   describe('getMapMarkers', () => {
@@ -47,7 +43,7 @@ describe(MapService.name, () => {
         country: asset.exifInfo!.country,
       };
       partnerMock.getAll.mockResolvedValue([]);
-      assetMock.getMapMarkers.mockResolvedValue([marker]);
+      mapMock.getMapMarkers.mockResolvedValue([marker]);
 
       const markers = await sut.getMapMarkers(authStub.user1, {});
