+++ server/test/repositories/map.repository.mock.ts
@@ -3,6 +3,9 @@ import { Mocked } from 'vitest';
 
 export const newMapRepositoryMock = (): Mocked<IMapRepository> => {
   return {
+    init: vitest.fn(),
+    reverseGeocode: vitest.fn(),
+    getMapMarkers: vitest.fn(),
     fetchStyle: vitest.fn(),
   };
 };
