+++ server/test/repositories/metadata.repository.mock.ts
@@ -3,9 +3,7 @@ import { Mocked, vitest } from 'vitest';
 
 export const newMetadataRepositoryMock = (): Mocked<IMetadataRepository> => {
   return {
-    init: vitest.fn(),
     teardown: vitest.fn(),
-    reverseGeocode: vitest.fn(),
     readTags: vitest.fn(),
     writeTags: vitest.fn(),
     extractBinaryTag: vitest.fn(),
